[Inductor Cutlass backend] Move tests to separate file
[ghstack-poisoned]
